webmvc/src/test/java/me/suhyuk/spring/webmvc/user/UserControllerTest.java
package me.suhyuk.spring.webmvc.user;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import static org.hamcrest.Matchers.equalTo;
import static org.hamcrest.Matchers.is;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@RunWith(SpringRunner.class)
// 우리는 slicing 기능 측, 웹만 사용할 것이므로
@WebMvcTest(UserController.class)
public class UserControllerTest {

    // HTTP 서버를 띄우지 않고 목업을 통해서 웹 액세스를 하는 것 처럼 동작하게 합니다
    @Autowired
    private MockMvc mockMvc;

    @Test
    public void testHello() throws Exception {
        mockMvc.perform(get("/hello"))
                .andExpect(status().isOk())
                .andExpect(content().string("hello"));
    }

    @Test
    public void testCreateUser() throws Exception {
        String userJson = "{\"username\":\"suhyuk\", \"password\":\"park\"}";
        mockMvc.perform(
                post("/users/create")
                .contentType(MediaType.APPLICATION_JSON_UTF8)
                .accept(MediaType.APPLICATION_JSON_UTF8)
                .content(userJson))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.username", is(equalTo("suhyuk"))))
                .andExpect(jsonPath("$.password", is(equalTo("park"))));
    }
}
